% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl7.R
\name{l7}
\alias{l7}
\title{l7}
\usage{
l7(opt, width = NULL, height = NULL)
}
\arguments{
\item{opt}{L7Plot options build from plot function with prefix 'l7_' and 'l7_opt()'}

\item{width}{optional number default: 400. width of the chart.}

\item{height}{optional number default: 400. height of the chart.}
}
\description{
l7() render a L7Plot
}
\examples{
l7_liquid(0.6) |>
  l7_opt(pattern=list(type='line')) |>
  l7()

l7_scatter(iris,'Sepal.Length','Sepal.Width',colorField='Species') |>
  l7_opt(
    legend=list(position='right'),
    shape='circle'
 ) |>
 l7()

}
\seealso{
Other l7: 
\code{\link{l7_dot}()}
}
\concept{l7}
