% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl7.R
\name{l7_choropleth}
\alias{l7_choropleth}
\title{l7_choropleth}
\usage{
l7_choropleth(map, source)
}
\arguments{
\item{map}{map}

\item{sourcesource}{}
}
\description{
l7_choropleth
}
\examples{
l7_choropleth(
  map= list(
    type='map',
    style='light',
    center=c(120.19382669582967, 30.258134),
    zoom=3,
    pitch=0
  ),
  source=list(
    data=data.frame(name=c('广东省','海南省'), value=c(100,50)),
    joinBy=list(
      sourceField='name',
      geoField='name'
    )
  )
  ) |>
    l7_opt(
      viewLevel=list(
        level='country',
        adcode=100000
      ),
      autoFit=TRUE,
      color=list(
        field='value',
        value=c('#B8E1FF', '#7DAAFF', '#3D76DD', '#0047A5', '#001D70'),
        scale=list(type='quantile')
      ),
      style=list(
        opacity=1,
        stroke='#ccc',
        lineWidth=0.6,
        lineOpacity=1
      ),
      label=list(
        visible=TRUE,
        field='name'
      )
    ) |> l7()
}
