% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl7.R
\name{l7_dot}
\alias{l7_dot}
\title{l7_dot}
\usage{
l7_dot(map, source)
}
\description{
opt param in rl7 is the PlotOptions for L7Plot.
You can construct an opt using function with \code{l7_} prefix and \code{l7_opt()}.
\url{https://l7plot.antv.antgroup.com/en/api} for details
}
\examples{

l7_scatter(mtcars, 'drat', 'wt') |>
  l7_opt(shape='circle', theme='dark') |>
  l7()


}
\seealso{
Other l7: 
\code{\link{l7}()}
}
\concept{l7}
