% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{l7_opt}
\alias{l7_opt}
\title{l7_opt}
\usage{
l7_opt(opt, ...)
}
\arguments{
\item{opt}{L7Plot options}

\item{...}{more options}
}
\description{
l7_opt
}
\examples{
l7_opt(list(), map=list(map="mapbox", zoom=3, pitch=0), source=list(...))


}
