% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{l7_output}
\alias{l7_output}
\title{Wrapper functions for using rl7 in shiny}
\usage{
l7_output(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{the width and height of the chart}
}
\description{
Use \code{l7/-output()} to create a UI element, and \code{render_l7()}
to render the map widget.
}
